cian_parser/pyproject.toml
[tool.poetry]
name = "cian-parser"
version = "0.1.0"
description = ""
authors = ["sanonesan <ezcudze@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
undetected-chromedriver = "^3.5.5"
seleniumbase = "^4.30.2"
configparser = "^7.1.0"
webdriver-manager = "^4.0.2"
pandas = "^2.2.2"
beautifulsoup4 = "^4.12.3"
tqdm = "^4.66.5"
loguru = "^0.7.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[virtualenvs]
prefer-active-python = true
